% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb_get_currencies.R
\name{get_all_currencies}
\alias{get_all_currencies}
\title{All currency values}
\usage{
get_all_currencies(date)
}
\arguments{
\item{date}{reference date}
}
\value{
A \code{data.frame} with all currency values from the given date.
The currency rates come quoted in BRL.
}
\description{
Gets all currency values
}
\examples{
\dontrun{
get_all_currencies("2017-03-10")
}
}
