% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb_get_market_expectations.R
\name{indics}
\alias{indics}
\title{Get available indicators for all APIs}
\usage{
indics(
  type = c("annual", "quarterly", "monthly", "inflation-12-months", "top5s-monthly",
    "top5s-annual", "institutions")
)
}
\arguments{
\item{type}{a character with one of the following: \code{annual}, \code{quarterly}
\code{monthly}, \code{inflation-12-months}, \code{top5s-monthly}, \code{top5s-annual},
\code{institutions}.

\code{type} defines the API used to fetch data.

\itemize{
\item \code{annual}: refers to the API *Expectativas de Mercado Anuais* for annual market expectations
\item \code{quarterly}: refers to the API *Expectativas de Mercado Trimestrais* for quarterly market expectations
\item \code{monthly}: refers to the API *Expectativas de Mercado Mensais* for monthly market expectations
\item \code{inflation-12-months}: refers to the API *Expectativas de mercado para inflação nos próximos 12 meses* for market expectations of inflation indexes for the next 12 months.
\item \code{top5s-monthly}: refers to the API *Expectativas de mercado mensais para os indicadores do Top 5* for monthly market expectations of top 5 indicators
\item \code{top5s-annual}: refers to the API *Expectativas de mercado anuais para os indicadores do Top 5* for annual market expectations of top 5 indicators
\item \code{institutions}: refers to the API *Expectativas de mercado informadas pelas instituições credenciadas* for market expectations sent by institutions
}}
}
\value{
A \code{character} vector with available indicators.
}
\description{
The best way to find out avaliable indicators for each endpoint.
}
\details{
Check <https://olinda.bcb.gov.br/olinda/servico/Expectativas/versao/v1/documentacao>
for more details
}
\examples{
\dontrun{
x <- indics("annual")
x <- indics("monthly")
}

}
