% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb_get_currencies.R
\name{list_currencies}
\alias{list_currencies}
\title{List all currencies}
\usage{
list_currencies()
}
\value{
A \code{data.frame} with information of all currencies
}
\description{
Lists all currencies and presents their name, symbol, numeric code, country name and county numeric code
}
\examples{
list_currencies()

}
