% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_ptax.R
\name{olinda_list_currencies}
\alias{olinda_list_currencies}
\title{List all currencies}
\usage{
olinda_list_currencies()
}
\value{
A \code{data.frame} with information of all currencies.

The \code{currency_type} refers to the currency's parity quotation.
Parity quotations relates currency values with USD.
}
\description{
Lists all currencies and presents their name, symbol, numeric code, country name and county numeric code
}
\examples{
\dontrun{
list_currencies()
}

}
