% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbcb-options}
\alias{rbcb-options}
\title{rbcb options}
\description{
Options used in rbcb inside some of its functions.
}
\details{
* `rbcb_cache`: all downloaded data is stored in temporary directories,
  if `rbcb_cache` is FALSE downloaded data overwrites files if it already
  exists. Otherwise, download is not executed and the existing file is
  returned. Defaults to TRUE.
* `rbcb_verbose`: if TRUE verbose messages are displayed when http requests
  are executed with httr. Defaults to FALSE.
}
\examples{
\dontrun{
options(rbcb_cache = FALSE)
options(rbcb_verbose = TRUE)
}
}
