% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb-package.R
\docType{package}
\name{rbcb-package}
\alias{rbcb-package}
\title{R interface to Brazilian Central Bank RESTful API}
\description{
The Brazilian Central Bank API delivers many datasets which regard economic
activity, regional economy, international economy, public finances, credit
indicators and many more.
These datasets can be accessed through \code{rbcb} functions and can be obtained in
different data structures common to R (tibble, data.frame, xts, ...).
}
