% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{rbcb_dataset}
\alias{rbcb_dataset}
\title{rbcb dataset}
\usage{
rbcb_dataset(name)
}
\arguments{
\item{name}{dataset name}
}
\value{
Addin has no return
}
\description{
`rbcb_dataset(name)` opens an [RStudio
  gadget](https://shiny.rstudio.com/articles/gadgets.html) and
  [addin](http://rstudio.github.io/rstudioaddins/) that allows you to view
  a few attributes that help to explain the desired data.
}
