% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb_addin.R
\name{rbcb_search}
\alias{rbcb_search}
\title{rbcb Search}
\usage{
rbcb_search(text = "")
}
\description{
`rbcb_search()` opens an [RStudio
  gadget](https://shiny.rstudio.com/articles/gadgets.html) and
  [addin](http://rstudio.github.io/rstudioaddins/) that allows you to query
  for specific terms and see a suitable rbcb command to fetch the
  desired data.
}
