% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb_search.R
\name{search_series}
\alias{search_series}
\title{Search for series in dadosabertos.bcb.gov.br site}
\usage{
search_series(q, page = 1)
}
\arguments{
\item{q}{query string}

\item{page}{page number to visualize of returning search

The \code{page} argument defaults to 1 which shows the first page of a returning search.
The search results also show the number of pages and the other pages can be showed by changing
this argument to the page to be presented.}
}
\value{
Display the search results on the screen
}
\description{
Search for series in dadosabertos.bcb.gov.br site
}
\examples{
search_series("IPCA")

}
