% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs.R
\name{sgs}
\alias{sgs}
\title{Create SGS code}
\usage{
sgs(..., load_info = TRUE)
}
\arguments{
\item{...}{numeric codes (preferably named)}

\item{load_info}{`logical` indicating with the dataset info shoud be loaded
(default TRUE)}
}
\value{
an SGS object representing SGS codes
}
\description{
SGS code is an objects that represents the SGS code used to download
datasets from the SGS API.
}
\examples{
\dontrun{
sgs(USD = 1, IPCA = 433)
}
}
