% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcb-package.R
\docType{package}
\name{rbcb-package}
\alias{rbcb}
\alias{rbcb-package}
\title{rbcb: R Interface to Brazilian Central Bank Web Services}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Brazilian Central Bank API delivers many datasets which regard economic activity, regional economy, international economy, public finances, credit indicators and many more. For more information please see \url{http://dadosabertos.bcb.gov.br/}. These datasets can be accessed through 'rbcb' functions and can be obtained in different data structures common to R ('tibble', 'data.frame', 'xts', ...).

The Brazilian Central Bank API delivers many datasets which regard economic
activity, regional economy, international economy, public finances, credit
indicators and many more.
These datasets can be accessed through \code{rbcb} functions and can be obtained in
different data structures common to R (tibble, data.frame, xts, ...).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilsonfreitas/rbcb}
  \item \url{https://wilsonfreitas.github.io/rbcb/}
  \item Report bugs at \url{https://github.com/wilsonfreitas/rbcb/issues}
}

}
\keyword{internal}
